(40 - 0.10.0.0) AUDIO STIMULI REFACTOR.  Audio mainly now uses AudioTrack API, not MediaPlayer. SoundPool was removed
audio stimuli are WAV encoded in 16-BIT PCM with the correct duration (10,20,30,35,50,100,1000 ms), sampled at 48KHz and saved in assets as WAV.
both MediaPlayer and AudioTrack autonomously detect their stop and prepare for next playback (no need to timer to call stop)

system delays are set according to the device
Tests' end management improved. added method to invoke resources release after test completion

TFI test now runs much more accurately

+++ psysuitecore/src/main/java/iit/uvip/psysuite/core/tests/sample/SubjectSampleDialogFragment.kt
@@ -271,7 +271,7 @@ open class SubjectSampleDialogFragment: SubjectBasicDialogFragment(), AdapterVie
             }
             (subject as SubjectSampleParcel).audioDuration   = etDurationAudio.text.toString().toLong()
             (subject as SubjectSampleParcel).audioResource   = spAudioResource.selectedItem as String
-            (subject as SubjectSampleParcel).audioVolume     = etAudioVolume.text.toString().toFloat()
+            (subject as SubjectSampleParcel).audioVolume     = etAudioVolume.text.toString().toInt()
         }
 
         if(swTactile.isChecked) {
